requirements.prod.txt
Django==5.2.7
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
gunicorn==22.0.0
dj-database-url==2.2.0
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.7
django-celery-beat==2.6.0
aiogram==3.6.0
aiohttp==3.9.5
whitenoise==6.7.0
